Backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
python-multipart
sqlalchemy
psycopg2-binary
geopandas
fiona
shapely
pyogrio
requests
